requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
reportlab==4.0.3
Pillow==11.0.0
